requirements.txt
Flask==3.1.2
pymongo==4.15.3
python-dotenv==1.2.1
colorama==0.4.6
